refactor: restructure project architecture for better organization
- Moved HomePage from components/pages to pages/ directory
- Centralized all tests in src/__tests__/ with mirrored structure
- Updated HomePage to use lazy loading for consistency
- Fixed all import paths after restructuring
- Added ARCHITECTURE.md documentation
- Improved separation of concerns between pages and components

Benefits:
- Clear separation between pages (routes) and reusable components
- Consistent lazy loading for all pages
- Centralized test organization
- Better scalability for future pages
